Validate client requests in corresponding directory ops
