+++ keyserver/listener.go
@@ -92,30 +92,19 @@ func handleOps(d *ConiksDirectory, req *Request) (*Response, ErrorCode) {
 	switch req.Type {
 	case RegistrationType:
 		if msg, ok := req.Request.(*RegistrationRequest); ok {
-			if len(msg.Username) > 0 && len(msg.Key) > 0 {
-				return d.Register(msg)
-			}
+			return d.Register(msg)
 		}
 	case KeyLookupType:
 		if msg, ok := req.Request.(*KeyLookupRequest); ok {
-			if len(msg.Username) > 0 {
-				return d.KeyLookup(msg)
-			}
+			return d.KeyLookup(msg)
 		}
 	case KeyLookupInEpochType:
 		if msg, ok := req.Request.(*KeyLookupInEpochRequest); ok {
-			if len(msg.Username) > 0 &&
-				msg.Epoch <= d.LatestSTR().Epoch {
-				return d.KeyLookupInEpoch(msg)
-			}
+			return d.KeyLookupInEpoch(msg)
 		}
 	case MonitoringType:
 		if msg, ok := req.Request.(*MonitoringRequest); ok {
-			if len(msg.Username) > 0 &&
-				msg.StartEpoch <= d.LatestSTR().Epoch &&
-				msg.StartEpoch <= msg.EndEpoch {
-				return d.Monitor(msg)
-			}
+			return d.Monitor(msg)
 		}
 	}
 	return NewErrorResponse(ErrorMalformedClientMessage),
